[COMMON] Faster aafConfig template
"index" function is bad in term of performance in Helm.
Reworked the templates in order to avoid it.

as certificates are retrieved at every boot (and as already present
certs are deleted before), we don't need persistent storage

Also set aafImage as a global variable in order to have a consistent use
accross ONAP

Issue-ID: EXTAPI-375
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: Ie3f5ae5c2a37d816afc42d2c67ebe8e40e749c79