Update patch set 1
Patch Set 1:

"index" was requested by Krzystof, I cannot judge.
persistence: 
Fine with me. 
For some reasons aaf-hello example is using persistence
I interpreted this as a robustness option if aaf is not available and pods are restarting and should continue working.
Or'aaf' is not used at all and persistence volume can be used as an alternative way to provide certs.
BR, Alex

Patch-set: 1