Update patch set 1
Patch Set 1:

Hi Alex,
It's not a "blame and shame" on you :). I just wanted to use your template and found some minor stuff to do!

> Patch Set 1:
> 
> "index" was requested by Krzystof, I cannot judge.

Krzysztof made a deep look at helm because we had this awful linting time and discovered that on tuesday so not your fault :)

> persistence: 
> Fine with me. 
> For some reasons aaf-hello example is using persistence
> I interpreted this as a robustness option if aaf is not available and pods are restarting and should continue working.
> Or'aaf' is not used at all and persistence volume can be used as an alternative way to provide certs.

That's what I thought but actually the script first destroy the certificates and __after__ tries to retrieve. So even when AAF is not working it won't be good...
Badly designed system...


> BR, Alex

Patch-set: 1