Update patch set 3
Patch Set 3:

The DCAE healthcheck is failing because 3 components (prh, the first-gen tca, and the non-tls ves collector) aren't coming up.  The bootstrap log is truncated, so we can't see what happened exactly.

The fact that we do not see Kubernetes pods for these components (they would have the "dep-" prefix) is a clue.  The k8s plugin responsible for managing the deployment uses the k8s API to create a k8s deployment.  It then waits for up to 1800 seconds for the deployment to become ready.  If the deployment is not ready, the plugin deletes the k8s deployment.  So possibly these components did not come up within 1800 seconds, possibly just because of resource limitations in the environment.

Patch-set: 3