Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> yeah I'm truncating all the logs because some takes more than 100M, don't ask me what we can do with that...
> 
> I'll see tomorrow european time how I can make an exception for dcae bootstrap container

If we could get the end of the log, rather than the beginning, I think we'd have the information we need.  (Of course, as soon as we start saving the end of the log, rather than the beginning,we'll encounter a problem where the beginning is what we need. :) )

Patch-set: 3
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0