[COMMON] Allow to include filePaths as a reference to variable
filePaths comes as a list and we didn't "fully support" passing this
variable as a reference to other variable like we do in all other
cases.

Let's fix that and allow both constructs:

secrets:
  - name: construct 1
    type: generic
    filePaths:
      - file1
      - file2
  - name: construct 2
    type: generic
    filePaths: '{{ .Values.fpaths }}'

fpaths: |
  - file1
  - file2
  - '{{ include "templateThatGeneratesFileName" . }}

Please note the | after : in fpaths. It means that from yaml point of
view this is is a string. We need to do it this way because we pass
this to tpl function and then we need to collect a proper list from it.

Issue-ID: SO-2730
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I5a6b475366bfea4cd0995a7e530bf88cb8ad639e