Update patch set 4
Patch Set 4:

This is a repeat of an earlier failure.  We're seeing none of the service components that are supposed to be deployed via Cloudify Manager at the request of the bootstrap container.  The bootstrap logs are truncated, so we cannot see why this happened.  (It's possible that the deployments were attempted but did not become ready within 30 minutes.  Or it's possible that, for some reason, they were never attempted.)

Patch-set: 4
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0