Update patch set 1
Patch Set 1:

It looks like the tcagen2 service was deployed but failed to become ready within 30 minutes.  It's not related to the change here, but it's not clear why this happened.  (When a service does not become ready within 30 minutes, the plugin deletes the k8s Deployment and the k8s Service(s) associated with it, which has the unfortunate side effect of deleting the logs.)

Patch-set: 1
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>