Update patch set 1
Patch Set 1:

We only recently added the healthcheck test for tcagen2, and it did not appear in OOM until change 104770, merged yesterday.  If you are seeing this a lot in the gating environment, it's possible that tcagen2 is slow to become ready or that it has some intermittent problem that we have not seen in our environments.

Patch-set: 1
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>