Update patch set 30
Patch Set 30: Code-Review-2

> Patch Set 30:
> 
> I agree. This is redundant, but was an intermediate step in the transition path for us to remove our dependency from sms, since sms is not seeing much traction and support in the community. To completely decouple our dependency from sms would require significant code changes which we couldn't pull off in this release.  

Ok it's cool that you would like to remove your sms dependency but this password encryption doesn't help with it and what it does instead it makes all those passwords extremely hard to change at the deployment time because the deployer needs to know:
- encryption algorithm
- exact tool that was used (as padding may be different)
- api key

So as you can see it starts being very annoying for someone who would like to deploy this in a production.

I expect that you encrypted those passwords because you don't want them to be stored on a disk in plain text,
The good news is that you don't need to encrypt them to ensure this. It would be enough if you use kubernetes secret and envsubst container that will process your config file and store the result only  using in memory volume.

That's why I recommend that you remove the encryption, otherwise I view those as "new hardcoded passwords" which are prohibited since Frankfurt

> 
> > Patch Set 30: Code-Review-1
> > 
> > > Patch Set 29:
> > > 
> > > (3 comments)
> > 
> > Ok. So now tell me what's the reason to encrypt passwords that you store in the vault? It doesn't make any sense to me...
> > 
> > Vault is considered to be the secret store so it should take and release plain text passwords and then use own security mechanisms to store the securely...

Patch-set: 30
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2