Update patch set 2
Patch Set 2:

@Sylvain, I have added more information under the jira ticket. 
Basically this is related to the email I sent earlier where we are looking forward to standardize the docs across projects to be able to add more automation to it. At first, it will require mass changes like this one, but this is a step needed in order for us to have more consistency across everyone. You can find more information here:
https://wiki.onap.org/display/DW/Docs+migration+tracker?src=contextnavpagetreemode

Once we are consistent using matching labels, we can start applying automated scripts to fix all broken linkages with minimal work.

Patch-set: 2
CC: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>