Update patch set 7
Patch Set 7:

Ehhh... I will try to debug this but I already started to hate policy module after fight with brmsgw. First it turned out that there was a typo since 2018, then it turned out that it was never possible to set a password to the DB and it worked only because the hardcoded default in the code matched hardcoded password... and obviously there is jo useful logs in stdout... ehhh... let's see what I will discover at this time.

Patch-set: 7