Update patch set 3
Patch Set 3: Code-Review-1

Sylvain, I'm not sure to get the point here, the superuser is always required to setup the DB and set a clamp user to access the clamp database.

This is already the case for Clamp, the mariadb setup is done by root superuser (the SQL is loaded by this user when the container starts), then clamp uses that "clds" user.
So what's the point here ?

Patch-set: 3
Reviewer: Gerrit User 23 <23@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1