Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> > Patch Set 3: Code-Review-1
> > 
> > Sylvain, I'm not sure to get the point here, the superuser is always required to setup the DB and set a clamp user to access the clamp database.
> > 
> > This is already the case for Clamp, the mariadb setup is done by root superuser (the SQL is loaded by this user when the container starts), then clamp uses that "clds" user.
> > So what's the point here ?
> 
> Here I see a "root" password set for an user named 'datarouter' (on patchset 2).
> So I assumed that "datarouter" user was used for everything.
> 
> If you have a specific user (if I understand well, clds) for clamp to access the DB, I think it would be good to offload login/password for this user that you have hardcoded (let's say here: https://github.com/onap/oom/blob/master/kubernetes/clamp/charts/mariadb/resources/config/mariadb/docker-entrypoint-initdb.d/bulkload/create-db.sql#L9 ;-) ) to values.yaml + secrets (see NBI for example)

OK then I see where we want to go, hardcoded is not really the right word, as everything is configurable here (even the SQL file that we provide), the problem is that you need to modify different files here (the sql and the values.yaml to provide the user cred to clamp).
What you want is more centralizing the creds or make it easier to specify the db access.
Also you want the creds to be auto-generated right ?
So indeed this change is not good, it just gives a name to the root user, which is obviously not what we want.

Patch-set: 3
Reviewer: Gerrit User 23 <23@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>