[COMMON] Allow to use ' in mysql passwords
derivePassword which we use to generate our passwords includes ' in
set of special characters that can be used in passwords.

Current implementation of bitnami configure-mysql.sh simply
concatenates password surrounded with '' rest of SQL query. This
causes issues if password contains ' as it creates invalid SQL statement.

To fix this issue we just patch the script and escape the special '
character in password.

Issue-ID: OOM-2246
Reported-by: Mateusz Pilat <m.pilat@partner.samsung.com>
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I3d2150760755e55558e2045cbb7ca5693c8ab71f