[COMMON] Add pre upgradre script for network name gen
Network-name-gen rolling upgrade needs to be supported to enable the
rolling upgrade of SDNC component. However, there is a value in values.yaml
when being true, changes the use of local nengdb to centralized mariadb-galera.
This involves in data migration and network name gen disruption during upgrade.
If this value is false, the new release continues to use the local nengdb and
rolling upgrade is feasible. We write a pre upgrade script to cater these two
events.

Issue-ID: OOM-2377
Change-Id: Ie3ca5d2fad940fcc8ea6e157430d4a90cd93b2b6
Signed-off-by: tringuyen <tri.nguyen@tatacommunications.com>