Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> DCAE HC are failing sorry

The healthcheck is failing because there is some problem with dcae-cloudify-manager.  Unfortunately the relevant logs are not written to stderr/stdout, so cannot tell why.  From the dcae-bootstrap-manager log, it is clear that dcae-cloudify-manager was working properly until suddenly it began returning "500 Server Error" responses to API requests.

There is nothing I can change in this submission to try to make a fix.  I suspect it's a transient problem, since I had no issues in my local environment.   The only change here is to update certificates.  There is no change to any logic and no change to dcae-cloudify-manager (other than that it got a new TLS certificate, which worked properly for many API requests).

Patch-set: 1