Update patch set 1
Patch Set 1:

In this case, the dcae-servicechange-handler never got past the initialization stage because the dmaap-message-router did not become available.

It appears, though, that the bootstrap process is also encountering error responses from the dcae-cloudify-manager API, as it did in the previous gating deployment.   I cannot reproduce this in my local environment.  There are clearly other problems with the gating deployment, but they don't appear to be related to dcae-cloudify-manager in any way.

Patch-set: 1