Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> >  When cert is just in OOM repo it's easy to replace so I expect than if you go to production first thing that you do is to replace all hardcoded certificates with your own;)
> 
> Well and I think that any app that is about to go to production shouldn't have any hardcoded site specific data but have a deployment option to populate own certificates on setup time :> Facility to update those certificates on run-time should be also a feature of production ready appliance.

Fully agree with you. THe problem is that we are not even close to this. What we are trying to do is to slowly, step by step migrate into this direction and this is only the first step. We don't want to ask for too much at once as in that case usually nothings gets done;)

Patch-set: 3
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>