Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I totally agree with you on the first point. On the second point however, there is documentation in place explaining the process.
> > 
> > how do you recommend we put this logic below of encrypting the pwd in an init container?
> >        
> >        /**
> > 	 * Encrypt the text using the secret key in key.properties file
> > 	 * 
> > 	 * @param value
> > 	 * @return The encrypted string
> > 	 * @throws BadPaddingException
> > 	 * @throws CipherUtilException
> > 	 *             In case of issue with the encryption
> > 	 */
> > 	public static String encryptPKC(String value, String skey) throws CipherUtilException {
> > 		Cipher cipher = null;
> > 		byte[] iv = null, finalByte = null;
> > 
> > 		try {
> > 			cipher = Cipher.getInstance(ALGORYTHM_DETAILS, "SunJCE");
> > 
> > 			SecureRandom r = SecureRandom.getInstance("SHA1PRNG");
> > 			iv = new byte[BLOCK_SIZE / 8];
> > 			r.nextBytes(iv);
> > 			ivspec = new IvParameterSpec(iv);
> > 			cipher.init(Cipher.ENCRYPT_MODE, getSecretKeySpec(skey), ivspec);
> > 			finalByte = cipher.doFinal(value.getBytes());
> > 
> > 		} catch (Exception ex) {
> > 			logger.error("encrypt failed", ex);
> > 			throw new CipherUtilException(ex);
> > 		}
> > 		return Base64.encodeBase64String(ArrayUtils.addAll(iv, finalByte));
> > 	}
> 
> There are two ways that I can see.
> 
> 1) Add a simple main to this class and just call it from the command line (use your application container as an init container by just replacing the entrypoint in helm)
> 
> 2) Use openssl to achieve the same encryption as implemented here in java

Tried openssl with multiple combinations, the behavior is different from the above java code and so unable to decrypt the password correctly. Christophe suggested another mechanism to talk to AAF instead of basic auth in which case this may not be required anymore. 

Proposal From Chirstophe:
Wouldn’t it be easier to reuse the X509 identity, that portal now generates on startup, as a mean to identify itself to AAF ? (that’s how it is supposed to work right ?)
So we could remove that password, key and encryption altogether from portal and reconfigure the portal user in AAF to be able to do whatever it is meant to do for portal work ?

Patch-set: 1
Reviewer: Gerrit User 679 <679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0