Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> Tested the new ESR versions in Frankfurt environment without success.

I will need someone to fix the frankfurt version of these components.  I reported early in the Frankfurt release that there was not support for a new version this release and we were going to use the elalto 1.4 versions. Then integration raised the flag on root containers, and said that we had to have it, so I tried to accomodate it.  @Peng, can you fix the frankfurt branch?  And 1.5.1 is the proper container ID, I had to upversion it.

Patch-set: 1