Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2:
> 
> @Jack @Vijay
> Could you guys if the HC is fine?
> I see that dep-dcae-prh is unavailable but I've 0 idea if it is related to this patch or not

prh isn't related to this change.  (We are seeing some occasional problems with the deployments that we are doing through Cloudify, but only in the gating environment.  They're not happening all the time, and it's not the same issue, and we don't see them in our local environment nor in integration testing.  So, a little bit of a mystery.)

Thanks @Krzysztof for doing all these password-related changes!

Patch-set: 2
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1