Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I'm concerned with the timing of this change.  This would be fine for Guilin, but I don't think we should be making changes like this for Frankfurt this late in the release

Well to be honest it doesn't change anything from the functional perspective and after the rendering the final helm chart should look very similar but it will use our common template instead of custom code at the source side.

Main motivation behind this change is that we are working to remove hardcoded truststore from docker aaf_agent docker image. If every component has a custom code it's a nightmare to do this (the truststore would now come from OOM repo as a configmap) and if everyone uses same template it's a relatively simple change.

Patch-set: 2