Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1: Code-Review-1
> 
> Temporary -1. We need to resolve username/password with policy team as healthcheck is not a suitable username

Well, this is not a critical issue. And at this stage such change is problematic. This password is used in many demos. Change may result with necessity of additional development and test of all the demos that use policy in the last moment...

Patch-set: 1
Reviewer: Gerrit User 2701 <2701@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1