Update patch set 1
Patch Set 1:

We're after RC1 so I believe this is for Guilin. 

If for Frankfurt, I need more informations on why we should merge this at this very late point in time and TSC to agree.

If for Guilin:

* You must first resolve for whole SO the hardcoded certificates (see nbi for implementation example)
* the commit message (on OOM) **must** follow this form:

```
[NAME_OF_COMPONENT|DOC|COMMON|GENERIC] Meaningful title (from OOM side)

at least one sentence explaining the change done in this patch, cause and consequences and possibly more of course

Issue-ID: AS_WE_ARE_FORCED_BUT_MEANINGLESS
Change-ID: xxx
Sign-off: xxx
```

Commit message will be the last stuff that will stay with our code so it must clearly explain the changes, the "why" and the consequences.
If it change OOM behavior in any way, documentation **must** be also updated.

Starting Frankfurt branching, merge requests which are not following this pattern will not be merged.

Please read the following pages and follow the guidelines for writing
commit message contained therein.

  - http://bit.ly/goodcommitmessages
  - http://who-t.blogspot.com/2009/12/on-commit-messages.html
  - http://dep.debian.net/deps/dep3/

* logs must be (also) on stdout

Patch-set: 1
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>