Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > When I take a look at the logs, I'm not sure it's really working...
> > https://gating-results.onap.eu/results//106960-1/infrastructure-healthcheck/k8s/pod-onap-esr-server-5db84f4df6-dbw87-esr-server-logs.html
> 
> Judging from the return of the error message, the error should be returned by the AAI, that is, the call has reached the AAI through the MSB, and the AAI prompts that the access failed: "Node Not Found: No Node of type cloud-region found at: cloud-infrastructure / cloud- regions "

Do, do we need a new container?  I am still unclear what our path forward is.  As I understand it - I released 1.5.1 of esr-gui and 1.5.3 or esr-server.  According to Andreas, a test case that works with 1.4.0 of esr-gui does not work with 1.5.1. Are there pending changes in a 1.5.2 container that can be released that will fix that issue?

Patch-set: 1