Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @Adrian, any ideas?

Hi Sylvain, I sent you an email on this. I think the NBI to SO seem to be working, service instance id was returned, but the test script says it was not. However the service order status was not complete, percentProgress reported by SO was 0%. Did SO complete the BPMN and update its status?. If it did then the test script might need tweaking in terms of its polling interval to check for completion.

Patch-set: 1
Reviewer: Gerrit User 1823 <1823@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>