[CONTRIB] Use proper image for netbox provision job
netbox was using readiness image to run its job. This is not the
proper usage of this image and recently it started failing because
bash is no longer available in this image.

As the script does not containt bash-specific construcitons let's
change the image to curl one and just use whatever shell is there.

Issue-ID: OOM-2406
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I0cc99c6d8fe0dbe59ee982d255753d149fcdd3eb