Update patch set 5
Patch Set 5:

I have a counterpart for "frankfurt" branch at https://gerrit.onap.org/r/c/oom/+/108046 that I am putting out as well.  I am not sure if it's the right thing to do, or just wait till this approved and cherry picked, but decided to put it out due as well due to the incoming RC2 date.

Patch-set: 5