Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Healthdist is up again. 
> @morgan, any reason why pnf registrate fails?

argh you merged, I have no access to the logs anymore

but we have also an issue on frankfurt today https://gating-results.onap.eu/results/onap_daily_pod4_frankfurt-567030023-05-26-2020_02-05/smoke-usecases-robot/pnf-registrate/pnf-registrate/report.html#totals?critical
2/3 PASS with 1 FAIL on modeling part

the ipaddress-v4-oam field retrieved from AA&I is empty 13.13.13.13 was expected
AA&I answered properly but it seems that the IP provided through the VES is not used anymore
the VES request looks correct, it includes this IP
but it seems to be dropped when we try to get it when calling the AAI API to check that the registration is done.

I checked it was working on the frankfurt daily of the 20th - we got some turbulences during the ascension break - no results, yesterday it was not executed and today we got this error.

Note on the 20th the only change on AAI (service-model) was already there

@Krzysztof K, you confirm?

we got one PASS today on a patch (108039-1 - merged)
http://testresults.opnfv.org/onap/api/v1/results?case_name=pnf-registrate&last=10&criteria=PASS
the previous PASS are from the 22nd

Patch-set: 1
Reviewer: Gerrit User 4481 <4481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>