Fix readiness-check ready.py script for stateful set
Only check the status field readyReplicas instead of
currentReplicas and updatedReplicas. The value of currentReplicas
relates to the number of pods created with the first revision of
the stateful set. If a change is made to the stateful set and pods are restarted
then currentReplicas is no longer returned in the status.
Check JIRA for more information.

Issue-ID: OOM-2418
Signed-off-by: Sebastien Premont-Tendland <sebastien.premont@bell.ca>
Change-Id: If5dbc892b6d807e6ca681457f9874613fce7db04