Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2: -Code-Review
> 
> > Patch Set 2:
> > 
> > Gitlab Orange 1: 
> > Full results can be seen here: https://gating-results.onap.eu/results//108127-2/index.html
> 
> pnf registrate seems to be not working
> @Morgan, @Krzystof K, can you check?
> 
> Otherwise it's OK on my side

for pnf-registrate, there is no log because the test did not finish or maybe did even not start properly

when we do not have the robot logs, we can always have a look at the xtesting log
https://gating-results.onap.eu/results//108127-2/smoke-usecases-robot/pnf-registrate/xtesting.log

we can see that the test is initiated then nothing
it usually means an init issue, first components not reachable or too slow.
which makes sense as the DCAE Healthcheck is FAIL (500)
and pnf-registrate is relying on it.

So the issue is on DCAE here
a priori not connected to SO update..
all the DCAE pods were detected as Running when performing the onap-k8s test..

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1