Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> waiting for DMaaP team +1

@Sylvain
I'm not sure if you took a look what's the difference with the previous patchset that was not working so let me warn you to not make the same stupid mistake that I did. When you copy-paste the item in requirements.yaml make sure to update also the chart alias. Otherwise the chart is included 2 times which leads to enormous amount of errors that you don't ever expected:D

Patch-set: 3