Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3: Code-Review+1
> > > 
> > > do (or should) we need to do the same for the other components?
> > 
> > It would be awesome if you could do this for other components as it would help with k8s v1.16+ migration
> 
> Ok. The idea was to do all the back and forth for one component only.
> Now that this change it completed and stabilized, I suggest to have it merged and then we'll have another one to apply the same for other remaining policy services. POLICY-2594 is created for that.

I fully agree with you. Smaller patches are easier to debug and to review. I'm just giving the floor to @Sylvain to review this as he is our service template expert;)

> When is the K8s 1.16 migration planned for?

We hope to do this in G but there is quite a lot of work before this is possible so we will see if we make it.

Patch-set: 3
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>