Update patch set 3
Patch Set 3: Code-Review+2

First, thank you for the patch!
If you want to continue on doing this on policy, it would be awesome (and one patch per subcomponent is good as the review is small, as my brain ;) )
Fyi, I'm working on other templates in order to ease some work and VID gerrit is the "showroom" of them (https://gerrit.onap.org/r/c/oom/+/107749)
Notice that "common.podSecurityContext" and "common.containerSecurityContext" may lead to error as it prevents write on rootfs.

So feel free to use them on other patches or I'll do that later on.

Patch-set: 3
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2