Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > And here we go... BUM!
> > 
> > Less than two months since we merge:
> > 
> > https://gerrit.onap.org/r/c/oom/+/101311
> > 
> > I've been complaining about this solution multiple times but no one from LFN was willing to take those comments into account "because we do it this way in all other projects" and "those are static file that will never be updated". As @Sylvain wrote in the review:
> > "I also think this is a bad solution and I keep my "I todd you so" for when I'll blow. __BUT__ everybody has moved except us so I'll merge. See you in six months / one year when this solution will fail"
> > 
> > So here we are after not 6 but only 2 months with those "static files" and having ~100 reviews:
> > 
> > https://gerrit.onap.org/r/q/CIMAN-376
> > 
> > Just to change which XSS file should be used. Don't you think sth is really wrong here? and if we merge this what we do with Frankfurt? Cherry-pick and have another 100 reviews?
> > 
> > *I'm really looking forward to have some proper solution for this as it has just been proven that the one which has been force pushed doesn't work*
> 
> Hi Krzy, 
> 
> We all understood your frustration. I agree this migration has this part as a downside but the benefits are so much better compared to submodules. 

And what are the benefits compared to a simple makefile that would just wget those files from a single repository when someone wants to build the documentation or compared to building this in a docker container that would already contain those files?

> Changes to config files are not done too often and I am willing to take the workload whenever it happens, I have committed to it. These changes were done as part of a script and did not take too much of my time to make.

I know that it's easy to automate but it's not only the matter of yours time. This is a typical synchronization problem and sooner or later those files will start to diverge between projects and what then? Gigantic sync of 100 repos?

It's a bad practice to do it in this way. We always should have single source of truth, this is what they taught all of us in universities right?

> It will take me some time to +2 and merge all, correct. But is something that I can help doing meanwhile we continue to make things better by adapting things to my script.

So now we get to the conclusion that it's not only your time but also time of people who need to review them. Time and resources needed to execute gating/build of those patches etc... I'm really curious how many jenkins worker hours we wasted in total to merge all those 100 review with such a single line change...

> We can think of adding a direct push rather than gerrit changes if think changes are small enough for example (not an official solution, im just providing an example. I would appreciate if you and the community provide their examples too on how can all make things better for docs and for teams). 
> For now, I don't see why this should be -2. Can you please provide a justification on why is this not a good change right now?

It's not a good change because the design behind it is broken and violates one of software development principles (https://en.wikipedia.org/wiki/Single_source_of_truth). The assertion when we got the previous change presented to us was that those file are static and as this patch shows they are not. That's why I strongly believe that instead of making such change and going further with this broken design it should be fixed.

If we want ONAP to follow best software development practices why don't we follow them for the infrastructure?

> 
> As agreed in the TSC, this is not needed for Frankfurt and teams do not need to cherry-pick this change. This change will be good to have in G release however hence the reason why I am merging it now so that teams will inherit it when branching for G

Ok that's a good news thanks!

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>