Update patch set 1
Patch Set 1:

Hi Krzysztof, first of all there is no need for this tone. If we stop for a second and go back to when we met in Prague, where you were sitting front row, really supporting this migration and offered to use the OSA project as a pilot. Not sure what happened between January and now. You are saying that you have been complaining multiple times but I don’t remember anyone speaking up during meetings or so. What I have seen is more less rude comments in Gerrit. This is open source and as far as I know, it is not fair to just shit on someone’s idea without a counter proposal or at least give some ideas/thoughts. So either we meet and try to sort this out, in a friendly manner or you will accept this solution a be a bit more supportive.

Patch-set: 1
Reviewer: Gerrit User 4857 <4857@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>