Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Krzysztof, first of all there is no need for this tone.

My tone was never supposed to be rude. If it sound like that it was not intentional the only thing I wanted to emphasize is that "I told you" but no one was willing to listen then... (see both OSA and OOM reviews when this change was introduced)

> If we stop for a second and go back to when we met in Prague, where you were sitting front row, really supporting this migration and offered to use the OSA project as a pilot. Not sure what happened between January and now. 

The answer is nothing!:) I still fully support getting out of this complicated submodule layout that causes some errors but in the same time I'm not supporting replacing one issue with another one which we actually did.

> You are saying that you have been complaining multiple times but I don’t remember anyone speaking up during meetings or so. 

U asked multiple times questions in gerrit and they've been never taken into account because "we do it this way in other projects" so well is this the open source spirit that you talk about? You can even see that I've never gave +1 nor +2 when this change was introduced to OOM and the only reason why PTL merged this was because we've been the last project with this change and everyone else just merged it.

> What I have seen is more less rude comments in Gerrit.

My comments were never supposed to be rude but technical. If they sounded rude it's only my fault and I apologize for it.

> This is open source and as far as I know, it is not fair to just shit on someone’s idea without a counter proposal or at least give some ideas/thoughts.

You are right this is open source and we should focus on producing the best possible software thus I believe that *technical criticism* is important and should be welcomed. Saying that this design does not follow one of software development best practices is a technical argument and considering that we have the same file in almost 100 repos it's rather a fact that a feeling.

I provided several ideas earlier and also in this review (makefile that wgets those file, docker container etc). I'm not saying that any of those is perfect and I'm happy to take *any* opinion on those proposals but I want to at least have a discussion on this so that we can collectively choose technically the best solution without pushing changes with argument "we do this already in 200 repos so 100 more is not a big deal". If we want our code to be of a really high quality we cannot afford for such changes.

> So either we meet and try to sort this out, in a friendly manner or you will accept this solution a be a bit more supportive.

I'm happy to participate in any kind of discussion that is finally going to be focused on technical aspects and try to figure out together how which technical solution should be used instead of just proving the solution to be good because XX other projects merged it.

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>